Initial Commit
This is the initial project creation of the Bebop Game Engine. This commit consists solely of the initial project file and a single main.cpp file that will be used for some initial testing. The main.cpp will eventually be removed in future updates as the project expands and separate games can be created using the engine.