Initial Window Class
This update is to add the initial creation of a display window. The basic setup of the Bebop namespace is used to create a Window class under the Bebop::Graphics namespace that will allow a user the create a window with their desired specifications. The main function added a quick test of this window creation.