Sprite Class
This update is for the creation of the basic sprite class that handles the creation, including the loading of a sprite image, and the drawing of the image to the window. Some of the allegro initialization has been moved into the main class temporary, but will eventually be placed into an engine specific class. The main function also has a testing of the sprite. Additionally cleaned up files from the last push that should not have been pushed into the repository.