Animated Sprite Class
This update is for the creation of animated sprites class. This class is a derivative of the sprite class and has a few extra inputs required compared to the static sprite class, including the frame change rate and number of frames in a animation.

This class will eventually be expanded upon to handle animation. An example of an animation change would be a sprite going from walking animation to an attacking animation and these animations require different number of frames for each animation and this class will need to be expanded to handle this.

Additionally, there are very minor grammar fixes and standards fixes.