Scene Class
This update adds a scene class to the project. The scene class handles managing a particular scene, such as a screen the player is on. Currently it handles holding sprites that are within the scene such as updating and drawing the sprites when called. The window class now has a scene item so now that a user can just use the window object to handle window related work so that pieces aren’t separated out. Eventually the window class will update to allow pushing new scenes and adding multiple scenes at a time (such as the ability to have some scenes simulating information while the main scene simulates AND draws). Eventually, these classes (scene and window) should be separated and a engine class should be made to handle both of these separately.

The main class now handles this as an example. One will notice the window object is handling the majority of the work. One nice thing is a user can create a sprite, push it into the scene and handle the business end outside the engine which will still be updated on the drawing in the engine.

Additionally, a few minor formatting issues are resolved and documentation for changes have been updated within the code.