Particle Classes
This update adds particle objects that are simple shaped objects. There are several points that may be taken out to separate class types such as shapes being separated as they would be used in other instances. Additionally, several expansions are planned for the future such as pathing (e.g., straight lines, swirl, wave, etc) for movement of the particle. For now, this is the basis of the particle for the engine.

Additionally fixed a few missing documentations from previous updates.