Object Classes
This update adds the start of the object types used in the game engine. This includes rectangular and circular objects and the basic uses such as positioning, center point, and drawing capabilities. As part of this update the particle classes has been reduced down to a single particle class that is now requiring an object as a parameter, taking composition over inheritance for the particle class. There is no longer child particle classes because now the main particle class will handle based on the object type passed in.

Additionally, a there has been small documentation updates within the code.