RotationMatrix2D
This update adds the RotationMatrix2D class that handles rotation through a matrix. The user can change the angle at any time. The sin wave class has been updated to utilize this new matrix class for the rotation of the sin wave. The particle class has been updated to now take a speed (in pixels) for the object to move at allowing particles to move at different speeds.

Additionally, updated a few documentation issues such as missing or incorrect documentation.