Math Motion Update - Circular Motion
This update primarily adds the circular motion class. The user can enter the radius and RPS (rotation/second) of the desired object. Part of this update includes the addition of a base motion class that the circular motion and sin wave motion (previously just sin wave) classes now inherit and implement the required functionality of this interface.

Additionally, I have updated some wrong documentation and corrected a few documentation problems.

There is also a new MathConstants class to handle all the constants required in the math namespace.

Reconfigured some of the layout of the folder paths to better represent what is going on in the engine such as motion has its own filter under the math filter to separate this functionality.