Color Class
This update adds a color class that handles holding RGB values along with an Alpha (opacity) value for the color. This essentially is used to wrap around the allegro color class so a user does not directly call allegro libraries for color.

The object classes have been enhanced to now take in this new color value. There is an example in the main class showing both the particles now having color to their respective object types. Currently the rectangle is red and the circle is green with some opacity to it.