Vector Class
This update adds the vector class. It currently only has the ability to do operations with other vectors. There are operations to add, subtract, multiply, and divide against vectors. There are also operations to find the magnitude and the normalized values of the vector. The vector direction is based on the positive and negative values of a component. A positive X component means movement to the right, negative is movement to the left. A positive Y component means movement downwards, negative is movement upwards.

There is a test case in the main class to show of the creation of several vectors as well as the operation of adding, subtracting, multiplying, and dividing vectors.

There are also a few minor fixes to layout of a other classes.