Light Class
This update adds the beginning of the light class. Currently it just creates a light within a 360 degree radius with any passed in origin point, light radius, and color. It creates this light by iterating around 360 degrees and creating triangles using a gradient coloring scheme to start the base color at the origin and at the edge fade to black.

This class will in the future be expanded to crate lights in limited radiuses and detect if the walls to create dynamic lights.

Additionally, this light has been added to the scene class and a test in the main function of a blue light has been made.