Improved Lights
This update improves the lighting effects for the engine. The scene class now handles separating drawing the lights color and a shadow map. The colors of the lights are first drawn onto the scene and the shadow map at the ending of the drawing sequence. The shadow map allows drawing a darkened mask over the scene with lite up areas based on the light information.

The light class now has an additional intensity variable that separates the intensity of the color and the intensity of the actual lights itself.

The main function has been updated to show off these new changes.