Collision Checker Improvement
Imprved the collision checker to now hold a vector for rectangles to be checked collisions. Now a user can add rectangles into the vector and call a new function that checks all possible collisions between the rectangles in the list (without repeating checks). Currently there is no collisioin handling so a check will just display if the two rectangles have collided or not through the iostream. This will eventually be changed to actually handle the collision. Another future improvement is to add other shaped objects for collision checking.