Static Library
This update changes this project from an application to a static library, something that needs to happen sooner rather than later. A separate test project for the original main function has been created to use this library, but will currently not be added into Github. Eventually projects to demo the engine will become available. Also started adding a new Bebop class for engine initialization that was needed for the allegro aspects of the engine. Unnecessary files have been removed.