Main Engine Class Update
Made an update to the main Bebop engine class to include the window and scene classes. This will decouple the window and scene class, allowing for possible future expansion of these classes without the scene being dependant on the window. There are plans in the future to allow multiple scenes, with one being drawn on and others being tacked in the background.