Collision Checker Update
Changed the idea behind the collision checker from being a handler to straight checking different kinds of collisions. Rectangle and Circle collision have been added, meaning that when two of there object have been placed into the checker it will return true if there is collision and flase if there is not. The current checks are Rectangle-Rectangle, Rectangle-Circle, Circle-Circle