Collision Detection Line Segment Collision Improvement
Collision involving line segments have been improved. The collision detection for two line segments now allow a user to pass in float pointers to retrieve the intersection point as well as still returning a boolean if there even was an inteersection. Line segment and rectangle has also been improved with the same feature. Additionally, the line and rectangle collision now properly detects the closest side of the rectangle that intersection happens.