Updated Collision Checker To Be Classless
Updated the collision checker to no longer be a class. The reason being is all functionality were only utility function calls so a class did not make sense in this case. All non public methods were wrapped into an anonymous namespace as they are not intended to be used outside of this class so any attempt to use these methods will result in a warning/error by the compiler. This will be easier to use these function to check collisions in future updates for the engine and allow easy use by any external calls for collision checks.