Lighting Update
Update lighting so now that objects that are supposed to block lighting will now block lights for rectangular and circular objects. This update is fairly extensive touching the object, light and scene classes to make this work. Essentially the light class tracks object that are marked to block lights when added into the scene class. Upon a scene class update the lights are recalculated which could potentially be expensive in large scenes but som ideas to help are being thought of. On a test with a moving object there was no slow down in recalculations and was able to smoothly do the recalculation. Several small fixes have been made in addition to the lighting update.