Fast Lighting Check
Updated to produce a fast lighting check. There are currently two problems with this current implementation. One being that circles are currently not checked. Second being that there is some unknown bug when an object is moving the calculation has a difficult time accurately recalculating the blocking points. For instance a rectangle moving in a circle will have so rays pass through it somehow, but if changed to be static it will calculate the points correctly. These two problems will be addressed shortly, but I wanted to push the current working implementation.