Circular Light Blocking Fixed
Added updated checks for blocking circular objects for lighting. This update imporves on the previous check where instead of sweeping the light and doing checks there will be an initial check against the circular object and light. If the circular object could potentially collide with the light then the collision will find the two edges of the circular the light will hit and then sweep between the edges to find blocking points. This is much faster implementation than the original implementation.

Circular objects with motion have no problems like the rectangular objects that have motion. Light blocking is smooth for circular object so there is not problem there like rectangular moving objects currently do.