Scene Layering
Added a class for scene layers. The main scene class now holds a map of these layers using an integer as the key so that layers are ordered based on ascending integer order. Layers are updated and drawn in the ascending order. This new layering allows lights and objects at different levels of the scene. For example if there is a layer 0 with a light and object that blocks the light the obejct will block the light sources and the object will end up casting a shadow being it. If a layers 1 is created to be above with a light, then this light will end up lighting the objects in layer 0 (mixing colors if the light has colors). This will have the object from layer 0 lit up now because a light is above it.