Event System
Started work on an event system to track particular events in the engine. This update reduces the need for the end user to track updating the updating/drawing of engine processes by only requiring the user to call the engine update call. This should be done within a while loop for the life of the game. The update will handle hte call to draw the scene when appropriate. Currentyl the event system only handles timing out based on a 60 FPS timeline and allows users to press keyboard keys, but does nothing at the moment but display a key was pressed/released. This will be expanded further for tracking what keys are pressed and released for the user to retrieve for their game. There are other expansions that will be added such as mouse input as well.