Updated Event - Track Key Presses
Updated the event class to now track key presses and if they are currently being pressed or not. This simply is a boolean check where true is the key being pressed and false being the key not pressed. Also updated the main class to allow a user to request this status by passing in the keycode, an integer. This is based off the ALLEGRO 5 keycodes.