Light Improvement
Mostly fixed a defect in the lighting where lighting had some strange snapping to object behind it. What this means is if an object blocks ligth and behind it is another object, inappropriate shadows would appear between the objects. This mostly fixes this problem though there is still an edge case that needs to be reviewed and fixed.