Light Update - Defect Fix
Fixed a defect where overlapping retangle objects that block a light were not taking in consideration of the collision points of these rectangles. This was causing shadows to appear inappropriately when a section should be lit, but was not. This was only fixed for rectangle/rectangle objects currently, but there is still a known issue for rectangle/circle to be addressed. Additionally, it is known the light class code is turning into spaghetti code and needs to be refactored to be clear as all the fixes while appropriate are becoming messy to maintain.