Lighting Update
Improved the lighting class by cleaning up a lot fo the problems that were found and now appropriately handles at the momeny rectangles. Circle algorithms have currently been removed, but placement to handle this is present and will eventually be reconfigured for this case. Overall this update improves accuracy of lighting greatly, fixes the spagehtti code problem the lighting class was starting to have and keeps good framerate for several lights.