Scene Layering Update
Updated classes to handle scene layering more properly. There was a defect where a light in a lower layer that would light objects in upper layers. This change helps fix problems for sprites and particles that are above a light source, that should be blocking the light from coming through. This is done by drawing a tinted version of the sprite/particle onto the shadow layer at shadow drawing time for that particular layer. Additionally, added to the sprite class to make RGB 255,0,255 to be transparent, so that color will become transparent if used in a sprite.