Lights Update - Optimizing
Updated to take of some optimization for the light class. Lowered the amount fo checks needed which helps processing time greatly when a lot of rectangular light objects are overlapping. Instead of checking every edge of a recangle against every edge of another rectangle it now decides which edges to compare based on the light location.