Refactoring
Refactored many classes to now use the Vector2D class for coordinates of objects instead of using two floats. This change affects many classes. Additionally, updated the update of the main class to now return a boolean where if true the update was due to a timeout happening for the FPS check otherwise a return of false will Happen (E.g., a button was pressed, mouse moved, etc.)