Update Lights - Light Cones
Updates the light file to now allow more than just a circular light. A user can now sepcify a starting angle of a light and an offset. The offset will be angle clockwise and counterclockwise from the center angle where the light will occurr. So as an example, to get a complete circular light, choose any center angle (e.g., 0), with an offset of 180 (180 degree clockwise and 180 degrees counterclockwise from the center angle) and this will create a complete circular light. To create a cone one might user a center angle of 90 degree (point straight downard on the scene) and offset of 30 degrees, creating a light from 60 to 120 degrees.